+++ package.json
@@ -40,6 +40,7 @@
     "@metamask/eslint-config-nodejs": "^12.0.0",
     "@metamask/eslint-config-typescript": "^12.0.0",
     "@metamask/eth-json-rpc-provider": "^1.0.0",
+    "@metamask/utils": "^6.2.0",
     "@types/node": "^16.18.24",
     "@typescript-eslint/eslint-plugin": "^5.30.7",
     "@typescript-eslint/parser": "^5.30.7",
@@ -56,6 +57,7 @@
     "execa": "^5.0.0",
     "isomorphic-fetch": "^3.0.0",
     "json-rpc-engine": "^6.1.0",
+    "nock": "^13.0.7",
     "prettier": "^2.7.1",
     "prettier-plugin-packagejson": "^2.2.17",
     "rimraf": "^3.0.2",
